refactor(markdown): move useHeadingSkipWarning to hooks/ per convention (FEC-984)
Component-specific custom hooks belong in components/<comp>/hooks/use-*.ts
(see docs/file-type-guidelines/hooks.md and main-component.md: the main file
should contain only the component it owns), mirroring pagination/hooks. Move
the inline hook out of markdown.tsx into hooks/use-heading-skip-warning.ts and
drop the now-unused getHeadingLevels/findHeadingLevelSkips imports there.

No behavior change; 26 markdown spec+story tests pass.
